Start using specific commits instead of -SNAPSHOT
This seems to be a better solution for IDE compatibility if things get
cached in a less-working version. Thanks for finding this problem and
working towards a solution, BenMcLean; I hit the same problem you had
with IDE caching, and it wasn't long after you first mentioned it. This
commit also updates some code in TheTsar to use stretchable fonts.