Add basic demo project for GdxSetup
I think this should fit the general criteria in the code; it needs the
working directory to be set to assets/ to run the desktop launcher in
the IDE but should otherwise run out of the box (except on HTML). I
think, for HTML, the extension needs fixing on the GdxSetup side because
the setup tool doesn't download the sources jar for squidlib-util or add
it as a GWT inherit, but does download the binary jar. I only selected
SquidLib, and it correctly added SquidLib (for display) and RegExodus as
it needs for HTML, but not SquidLib-Util.