New version saves text between runs, on GWT too!
Because GWT doesn't have a good way to copy from the libGDX application,
this produces a big purple text field when run as a GWT app. When you
click generate, it fills it with all the stuff you might want to copy
out, including the source text, the ciphered text, and the special seed
this uses for the language. It might need Ctrl-A to select all text in
the field, then Ctrl-C to copy it; there's some weird bug that makes it
hard to select text here.