More unexpected results!
Testing nextLong() as well as nextInt() shows Zig32RNG is always faster
than LightRNG and RandomXS128, even when it's out of its element on
64-bit generation. It can't generate all longs, though, since it uses a
pair of ints to generate a long. Light32RNG isn't tested now because it
doesn't produce the same results on GWT and on desktop, so it's not
really reliable.