Include more PRNGs in benchmark, use SquidLib Zag
ThrustAlt32RNG is a little faster than Zag32RNG, but can't produce all
ints and has a wimpy period of 2 to the 32. XoRoRNG does better than
LightRNG on GWT, but not as well as RandomXS128 (likely due to an
inability to use the SSE bitwise rotation instructions on longs with
GWT, while they speed up XoRo on desktop relative to the xorshift128+
algorithm that RandomXS128 uses). Zag32RNG is in SquidLib now, so we
just depend on a more recent commit, with much gratitude to JitPack.io !