Speedup with Zug32RNG; waiting on quality tests
Zug32RNG is able to omit one of the six xorshift operations that
Zag32RNG uses by switching to an LCG instead of an LCG and adjusting the
shift amounts. This seems to show a small speed boost, approaching
ThrustAlt32RNG on nextInt() and exceeding anything else tested on
nextLong(). Quality tests are still going and should be for at least a
day more if they result in a successful pass; 8TB have been passed so
far and what had been rough areas on earlier tests have gone smoothly
this time, with two "unusual" anomalies and nothing worse than that.