RNG! Add XoRo32 (faster, lower quality), Zog32
Zug32 has also been updated with a version that passes PractRand with
only minor changes. XoRo32 is the fastest generator here, but fails
several quality tests, and not just the allegedly-irrelevant binary
matrix rank test. It is yet to be seen if it can have its quality
improved without harming the period or slowing it down very much.
Zog32RNG, on the other hand, only had one anomaly, and no failures, over
32TB of PractRand testing, but it isn't much different on speed relative
to Zag32RNG or Zug32RNG, so I'm keeping Zag as-is in SquidLib for now.
It's possible Zog can be sped up further, in which case it might replace
or augment Zag.