Add KeyRemap jar, tool used to minify it
Using Deconimus/JarShrink (with minor changes to get it to run on my JDK
setup), KeyRemapTool.jar went from 8.5 MB to about 5.9 MB (this needed
the latest SquidLib commit that reduces the length of the import
dependency chain that made KeyRemap hard to minify before). The slightly
changed JarShrink is included here now. KeyRemap has switched back to
LWJGL 2 because it results in a slightly smaller output JAR.