Improve NorthernLights; it doesn't stretch like taffy now
The repeated "stretching" of the colors in NorthernLights lasted longer than it should have and was a fairly boring part of the light show; it happened when a multiplier obtained from swayRandomized() was too close to 0 for too long. Now the swayRandomized() used here always returns a positive number and averages at 2.5 or so instead of 0, so any stretch time winds up being very brief.